[Api] setup.sh improvements

+++ attendance-api/setup.sh
@@ -1,11 +1,14 @@
-cp .env.example .env
+[ -f .env ] || cp .env.example .env
 
 pushd $(dirname $(readlink -f $0))
 
-docker run -it --rm -v ${PWD}/:/opt -w /opt laravelsail/php74-composer:latest bash -c "composer install && php artisan key:generate"
+docker run -it --rm -v ${PWD}/:/opt -w /opt laravelsail/php74-composer:latest bash -c "composer install"
 
 ./vendor/bin/sail up -d
 sleep 5
+
+./vendor/bin/sail artisan key:generate
+./vendor/bin/sail artisan migrate:refresh --seed
 
 ./vendor/bin/sail stop
